requirements.txt
pymongo[srv]~=4.2.0
-e .
from-root
certifi~=2022.9.24
setuptools~=63.4.1
python-dotenv~=0.21.0
pandas~=1.3.5
scikit-learn==1.0.2
numpy==1.21.6
dill
PyYAML
scipy
imblearn
xgboost
fastapi[all]
starlette